Refactor ExceptionHandlingMiddleware
