--- Slim/Interfaces/ExceptionHandlerInterface.php
@@ -1,23 +1,0 @@
-<?php
-
-/**
- * Slim Framework (https://slimframework.com)
- *
- * @license https://github.com/slimphp/Slim/blob/5.x/LICENSE.md (MIT License)
- */
-
-declare(strict_types=1);
-
-namespace Slim\Interfaces;
-
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Throwable;
-
-interface ExceptionHandlerInterface
-{
-    public function __invoke(
-        ServerRequestInterface $request,
-        Throwable $exception,
-    ): ResponseInterface;
-}
