+++ tests/Middleware/ExceptionHandlingMiddlewareTest.php
@@ -10,113 +10,39 @@ declare(strict_types=1);
 
 namespace Slim\Tests\Middleware;
 
+use DOMDocument;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
-use Psr\Http\Message\ResponseFactoryInterface;
-use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestFactoryInterface;
-use Psr\Http\Message\ServerRequestInterface;
 use RuntimeException;
 use Slim\Builder\AppBuilder;
-use Slim\Interfaces\ExceptionHandlerInterface;
+use Slim\Error\Renderers\HtmlExceptionRenderer;
+use Slim\Error\Renderers\JsonExceptionRenderer;
+use Slim\Error\Renderers\XmlExceptionRenderer;
+use Slim\Media\MediaType;
 use Slim\Middleware\EndpointMiddleware;
 use Slim\Middleware\ExceptionHandlingMiddleware;
 use Slim\Middleware\RoutingMiddleware;
 use Slim\Tests\Traits\AppTestTrait;
-use Throwable;
 
 final class ExceptionHandlingMiddlewareTest extends TestCase
 {
     use AppTestTrait;
 
-    public function testExceptionHandlingMiddlewareHandlesException()
+    public function testDefaultHandlerWithoutDetails(): void
     {
         $builder = new AppBuilder();
-        $app = $builder->build();
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
 
-        $responseFactory = $app->getContainer()->get(ResponseFactoryInterface::class);
-
-        // Custom ExceptionHandlerInterface implementation
-        $exceptionHandler = new class ($responseFactory) implements ExceptionHandlerInterface {
-            private ResponseFactoryInterface $responseFactory;
-
-            public function __construct($responseFactory)
-            {
-                $this->responseFactory = $responseFactory;
-            }
-
-            public function __invoke(ServerRequestInterface $request, Throwable $exception): ResponseInterface
-            {
-                $response = $this->responseFactory->createResponse(500, 'Internal Server Error');
-                $response->getBody()->write($exception->getMessage());
-
-                return $response;
-            }
-        };
-
-        $app->add((new ExceptionHandlingMiddleware())->withExceptionHandler($exceptionHandler));
-        $app->add(RoutingMiddleware::class);
-        $app->add(EndpointMiddleware::class);
-
-        $app->get('/', function () {
-            throw new RuntimeException('Something went wrong');
-        });
-
-        $request = $app->getContainer()
-            ->get(ServerRequestFactoryInterface::class)
-            ->createServerRequest('GET', '/');
-
-        $response = $app->handle($request);
-        $this->assertEquals(500, $response->getStatusCode());
-        $this->assertSame('Something went wrong', (string)$response->getBody());
-    }
-
-    public function testExceptionHandlingMiddlewarePassesThroughNonExceptionRequest()
-    {
-        $builder = new AppBuilder();
-        $app = $builder->build();
-
-        $responseFactory = $app->getContainer()->get(ResponseFactoryInterface::class);
-
-        // This handler should not be called in this test
-        $exceptionHandler = new class ($responseFactory) implements ExceptionHandlerInterface {
-            private ResponseFactoryInterface $responseFactory;
-
-            public function __construct($responseFactory)
-            {
-                $this->responseFactory = $responseFactory;
-            }
-
-            public function __invoke(ServerRequestInterface $request, Throwable $exception): ResponseInterface
-            {
-                $response = $this->responseFactory->createResponse(500);
-                $response->getBody()->write($exception->getMessage());
-
-                return $response;
-            }
-        };
-
-        $app->add(new ExceptionHandlingMiddleware($exceptionHandler));
-        $app->add(RoutingMiddleware::class);
-        $app->add(EndpointMiddleware::class);
-
-        $app->get('/', function (ServerRequestInterface $request, ResponseInterface $response) {
-            $response->getBody()->write('Hello World');
-
-            return $response;
-        });
-
-        $request = $app->getContainer()
-            ->get(ServerRequestFactoryInterface::class)
-            ->createServerRequest('GET', '/');
-
-        $response = $app->handle($request);
-        $this->assertEquals(200, $response->getStatusCode());
-        $this->assertSame('Hello World', (string)$response->getBody());
-    }
-
-    public function testDefaultMediaTypeWithoutDetails(): void
-    {
-        $builder = new AppBuilder();
+                    return $middleware
+                        ->withDisplayErrorDetails(false)
+                        ->withDefaultHandler(HtmlExceptionRenderer::class);
+                },
+            ]
+        );
         $app = $builder->build();
 
         $app->add(ExceptionHandlingMiddleware::class);
@@ -139,10 +65,20 @@ final class ExceptionHandlingMiddlewareTest extends TestCase
         $this->assertStringContainsString('<h1>Application Error</h1>', (string)$response->getBody());
     }
 
-    public function testDefaultHtmlMediaTypeWithDetails(): void
+    public function testDefaultHandlerWithDetails(): void
     {
         $builder = new AppBuilder();
-        $builder->addSettings(['display_error_details' => true]);
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware
+                        ->withDisplayErrorDetails(true)
+                        ->withDefaultHandler(HtmlExceptionRenderer::class);
+                },
+            ]
+        );
         $app = $builder->build();
 
         $app->add(ExceptionHandlingMiddleware::class);
@@ -165,10 +101,60 @@ final class ExceptionHandlingMiddlewareTest extends TestCase
         $this->assertStringContainsString('<h1>Application Error</h1>', (string)$response->getBody());
     }
 
-    public function testJsonMediaTypeWithDetails(): void
+    public function testDefaultHtmlMediaTypeWithDetails(): void
     {
         $builder = new AppBuilder();
-        $builder->addSettings(['display_error_details' => true]);
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware
+                        ->withDisplayErrorDetails(true)
+                        ->withDefaultMediaType(MediaType::TEXT_HTML)
+                        ->withHandler(MediaType::TEXT_HTML, HtmlExceptionRenderer::class);
+                },
+            ]
+        );
+        $app = $builder->build();
+
+        $app->add(ExceptionHandlingMiddleware::class);
+        $app->add(RoutingMiddleware::class);
+        $app->add(EndpointMiddleware::class);
+
+        $request = $app->getContainer()
+            ->get(ServerRequestFactoryInterface::class)
+            ->createServerRequest('GET', '/')
+            ->withHeader('Accept', 'application/json');
+
+        $app->get('/', function () {
+            throw new RuntimeException('Test error', 123);
+        });
+
+        $response = $app->handle($request);
+
+        $this->assertSame(500, $response->getStatusCode());
+        $this->assertSame('text/html', (string)$response->getHeaderLine('Content-Type'));
+        $this->assertStringNotContainsString('Test Error message', (string)$response->getBody());
+        $this->assertStringContainsString('<h1>Application Error</h1>', (string)$response->getBody());
+    }
+
+    public function testJsonMediaTypeDisplayErrorDetails(): void
+    {
+        $builder = new AppBuilder();
+
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware
+                        ->withDisplayErrorDetails(true)
+                        ->withHandler(MediaType::APPLICATION_JSON, JsonExceptionRenderer::class);
+                },
+            ]
+        );
+
         $app = $builder->build();
 
         $app->add(ExceptionHandlingMiddleware::class);
@@ -197,12 +183,12 @@ final class ExceptionHandlingMiddlewareTest extends TestCase
     public function testWithoutHandler(): void
     {
         $this->expectException(RuntimeException::class);
-        $this->expectExceptionMessage('Test error');
+        $this->expectExceptionMessage('Exception handler for "text/html" not found');
 
         $builder = new AppBuilder();
         $app = $builder->build();
 
-        $app->add(new ExceptionHandlingMiddleware());
+        $app->add(ExceptionHandlingMiddleware::class);
         $app->add(RoutingMiddleware::class);
         $app->add(EndpointMiddleware::class);
 
@@ -215,5 +201,156 @@ final class ExceptionHandlingMiddlewareTest extends TestCase
         });
 
         $app->handle($request);
+    }
+
+    #[DataProvider('textHmlHeaderProvider')]
+    public function testWithTextHtml(string $header, string $headerValue): void
+    {
+        $builder = new AppBuilder();
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware
+                        ->withDisplayErrorDetails(true)
+                        ->withHandler(MediaType::TEXT_HTML, HtmlExceptionRenderer::class);
+                },
+            ]
+        );
+        $app = $builder->build();
+
+        $app->add(ExceptionHandlingMiddleware::class);
+        $app->add(RoutingMiddleware::class);
+        $app->add(EndpointMiddleware::class);
+
+        $app->get('/', function () {
+            throw new RuntimeException('Test Error message');
+        });
+
+        $request = $app->getContainer()
+            ->get(ServerRequestFactoryInterface::class)
+            ->createServerRequest('GET', '/')
+            ->withHeader($header, $headerValue);
+
+        $response = $app->handle($request);
+
+        $this->assertSame(500, $response->getStatusCode());
+        $this->assertSame('text/html', (string)$response->getHeaderLine('Content-Type'));
+        $this->assertStringContainsString('Test Error message', (string)$response->getBody());
+    }
+
+    public static function textHmlHeaderProvider(): array
+    {
+        return [
+            ['Accept', 'text/html'],
+            ['Accept', 'text/html, application/xhtml+xml, application/xml;q=0.9, image/webp, */*;q=0.8'],
+            ['Content-Type', 'text/html'],
+            ['Content-Type', 'text/html; charset=utf-8'],
+        ];
+    }
+
+    // todo: Add test for other media types
+
+    public function testWithAcceptJson(): void
+    {
+        $builder = new AppBuilder();
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware
+                        ->withDisplayErrorDetails(false)
+                        ->withHandler(MediaType::APPLICATION_JSON, JsonExceptionRenderer::class);
+                },
+            ]
+        );
+        $app = $builder->build();
+
+        $app->add(ExceptionHandlingMiddleware::class);
+        $app->add(RoutingMiddleware::class);
+        $app->add(EndpointMiddleware::class);
+
+        $request = $app->getContainer()
+            ->get(ServerRequestFactoryInterface::class)
+            ->createServerRequest('GET', '/')
+            ->withHeader('Accept', 'application/json');
+
+        $app->get('/', function () {
+            throw new RuntimeException('Test exception');
+        });
+
+        $response = $app->handle($request);
+
+        $this->assertSame(500, $response->getStatusCode());
+        $expected = [
+            'message' => 'Application Error',
+        ];
+        $this->assertJsonResponse($expected, $response);
+    }
+
+    public static function xmlHeaderProvider(): array
+    {
+        return [
+            ['Accept', 'application/xml'],
+            ['Accept', 'application/xml, application/json'],
+            ['Content-Type', 'application/xml'],
+            ['Content-Type', 'application/xml; charset=utf-8'],
+        ];
+    }
+
+    #[DataProvider('xmlHeaderProvider')]
+    public function testWithAcceptXml(string $header, string $headerValue): void
+    {
+        $builder = new AppBuilder();
+        $builder->addDefinitions(
+            [
+                ExceptionHandlingMiddleware::class => function ($container) {
+                    $middleware = ExceptionHandlingMiddleware::createFromContainer($container);
+
+                    return $middleware->withDisplayErrorDetails(false)
+                        ->withoutHandlers()
+                        ->withHandler('application/json', JsonExceptionRenderer::class)
+                        ->withHandler('application/xml', XmlExceptionRenderer::class);
+                },
+            ]
+        );
+        $app = $builder->build();
+
+        $app->add(ExceptionHandlingMiddleware::class);
+        $app->add(RoutingMiddleware::class);
+        $app->add(EndpointMiddleware::class);
+
+        $request = $app->getContainer()
+            ->get(ServerRequestFactoryInterface::class)
+            ->createServerRequest('GET', '/')
+            ->withHeader($header, $headerValue);
+
+        $app->get('/', function () {
+            throw new RuntimeException('Test exception');
+        });
+
+        $response = $app->handle($request);
+
+        $this->assertSame(500, $response->getStatusCode());
+        $expected = '<?xml version="1.0" encoding="UTF-8"?>
+                    <error>
+                      <message>Application Error</message>
+                    </error>';
+
+        $dom = new DOMDocument();
+        $dom->preserveWhiteSpace = false;
+        $dom->formatOutput = true;
+        $dom->loadXML($expected);
+        $expected = $dom->saveXML();
+
+        $dom2 = new DOMDocument();
+        $dom2->preserveWhiteSpace = false;
+        $dom2->formatOutput = true;
+        $dom2->loadXML((string)$response->getBody());
+        $actual = $dom2->saveXML();
+
+        $this->assertSame($expected, $actual);
     }
 }
